shared/azure_clients.py
"""
Azure AI services client module.

Provides clients for Azure OpenAI (GPT-5) for AI agent functionality.
"""

import os
import logging
from typing import Optional
from azure.identity import DefaultAzureCredential

logger = logging.getLogger(__name__)


class AzureOpenAIClient:
    """Wrapper for Azure OpenAI client."""

    def __init__(self):
        """Initialize Azure OpenAI client with credentials from environment."""
        self.endpoint = os.getenv(
            "AZURE_OPENAI_ENDPOINT",
            "https://brn-azai.cognitiveservices.azure.com"
        )
        self.api_key = os.getenv("AZURE_OPENAI_API_KEY")
        self.deployment_name = os.getenv("GPT5_DEPLOYMENT_NAME", "gpt-5")
        self.api_version = "2025-01-01-preview"

        if not self.api_key:
            logger.warning("AZURE_OPENAI_API_KEY not set - Azure OpenAI will not be available")
            self.client = None
        else:
            logger.info(f"Azure OpenAI client configured for deployment: {self.deployment_name}")
            # Client initialization will be done when needed
            self.client = None

    def get_chat_completion(self, messages: list, temperature: float = 0.7, max_tokens: int = 1000):
        """
        Get chat completion from Azure OpenAI.

        Args:
            messages: List of message dictionaries with 'role' and 'content'
            temperature: Sampling temperature (0-1)
            max_tokens: Maximum tokens in response

        Returns:
            String response from the model

        Note: This is a placeholder for future AI agent implementation.
        """
        if not self.client:
            logger.warning("Azure OpenAI client not initialized")
            return None

        try:
            # Placeholder for actual implementation
            # This will be implemented when AI agent functionality is added
            logger.info("Chat completion requested (not yet implemented)")
            return None

        except Exception as e:
            logger.error(f"Failed to get chat completion: {e}")
            raise

    def analyze_market_sentiment(self, market_data: dict) -> dict:
        """
        Analyze market sentiment using GPT-5.

        Args:
            market_data: Market information dictionary

        Returns:
            Dictionary with sentiment analysis

        Note: This is a placeholder for future AI agent implementation.
        """
        logger.info("Market sentiment analysis requested (not yet implemented)")
        return {
            'sentiment': 'neutral',
            'confidence': 0.0,
            'reasoning': 'AI analysis not yet implemented'
        }


# Module-level singleton instance
_openai_client: Optional[AzureOpenAIClient] = None


def get_azure_openai_client() -> AzureOpenAIClient:
    """
    Get or create Azure OpenAI client singleton.

    Returns:
        AzureOpenAIClient: Client instance
    """
    global _openai_client

    if _openai_client is None:
        logger.info("Creating Azure OpenAI client instance")
        _openai_client = AzureOpenAIClient()

    return _openai_client
