requirements.txt
# Azure Functions
azure-functions==1.20.0
azure-identity==1.18.0

# Polymarket Integration
py-clob-client==0.28.0

# Azure OpenAI (for sentiment analysis)
openai==1.54.0

# Google Gemini (for fallback sentiment)
google-generativeai==0.8.3

# Database
psycopg2-binary==2.9.10

# HTTP & Utilities
requests==2.32.3
python-dotenv==1.0.1

# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python
# azure-monitor-opentelemetry
